Moving type out from being nested
