+++ src/AzureOpenAIProxy.Management/Components/EventManagement/EventEditor.razor.cs
@@ -1,4 +1,3 @@
-using System.ComponentModel.DataAnnotations;
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Forms;
 
@@ -52,39 +51,5 @@ public partial class EventEditor : ComponentBase
         isSubmitting = true;
         await OnValidSubmit.InvokeAsync(Model);
         isSubmitting = false;
-    }
-
-    public class EventEditorModel
-    {
-        [Required]
-        public string? Name { get; set; }
-
-        [Required(ErrorMessage = "Url text is required")]
-        [StringLength(256)]
-        public string? UrlText { get; set; }
-        [Required(ErrorMessage = "Url is required")]
-        [StringLength(256)]
-        [Url]
-        public string? Url { get; set; }
-        [Required]
-        public string? Description { get; set; }
-        [Required]
-        public DateTime? Start { get; set; }
-        [Required]
-        public DateTime? End { get; set; }
-        [Required(ErrorMessage = "Organizer name is required")]
-        [StringLength(128)]
-        public string? OrganizerName { get; set; }
-        [Required(ErrorMessage = "Organizer email is required")]
-        [StringLength(128)]
-        [EmailAddress]
-        public string? OrganizerEmail { get; set; }
-        [Required(ErrorMessage = "Specify the maximum number of tokens allowed per request")]
-        public int MaxTokenCap { get; set; } = 4096;
-        [Required]
-        public bool SingleCode { get; set; }
-        [Required(ErrorMessage = "Specify the maximum number of requests allowed per day per token")]
-        public int DailyRequestCap { get; set; } = 10000;
-        public bool Active { get; set; }
     }
 }
